requirements.txt
Flask==2.3.3
Flask-SQLAlchemy>=2.5.0,<3.1.0
SQLAlchemy>=1.4.0,<2.0.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
psycopg2-binary==2.9.7
bleach==6.0.0
MarkupSafe==2.1.3
python-dotenv==1.0.0
Pillow==10.0.1
Flask-CSRF==0.9.2
gunicorn==21.2.0
cloudinary==1.36.0 